Made Major CHanges to the CAN Loop
The pod is now running. To run the pod. FIrst run the Host.vi and then run then deploy the CRIO Code. After wait for the GUI to display the current state as idle. Once the Pod is currently in the Idle state. Enable Service, input the target rpm ( should be negative for the current wiring) After press drive enable. Select ControlWord 6 and Control Word 15. Then press and hold the drive button. To immediately stop the pod press return to idle